fix(middlewares): remove region check when querying orgs
