+++ apps/microsoft/src/database/client.ts
@@ -1,16 +1,11 @@
-import { sql } from '@vercel/postgres';
-import { drizzle } from 'drizzle-orm/vercel-postgres';
-import { neonConfig } from '@neondatabase/serverless';
+import { drizzle } from 'drizzle-orm/postgres-js';
+import postgres from 'postgres';
 import { env } from '@/env';
-import * as schema from './schema';
 
-if (!env.VERCEL_ENV || env.VERCEL_ENV === 'development') {
-  // Set the WebSocket proxy to work with the local instance
-  neonConfig.wsProxy = (host) => `${host}:${env.POSTGRES_PROXY_PORT}/v1`;
-  // Disable all authentication and encryption
-  neonConfig.useSecureWebSocket = false;
-  neonConfig.pipelineTLS = false;
-  neonConfig.pipelineConnect = false;
-}
+const queryClient = postgres(
+  env.VERCEL_ENV && env.VERCEL_ENV !== 'development'
+    ? `${env.POSTGRES_URL}?sslmode=require`
+    : env.POSTGRES_URL
+);
 
-export const db = drizzle(sql, { schema });
+export const db = drizzle(queryClient);
